very_basic_and_unorganised_static_stair_mech
